Some hierarchy to separate the implementations
