+++ VultisigApp/VultisigApp/Services/CoinService/CosmosCoinService.swift
@@ -1,0 +1,11 @@
+//
+//  CosmosCoinService.swift
+//  VultisigApp
+//
+//  Created by Enrique Souza Soares on 05/07/24.
+//
+
+import Foundation
+
+@MainActor
+class CosmosCoinService: CoinService {}
