+++ VultisigApp/VultisigApp/Services/CoinService/UtxoCoinService.swift
@@ -1,0 +1,11 @@
+//
+//  UtxoCoinService.swift
+//  VultisigApp
+//
+//  Created by Enrique Souza Soares on 05/07/24.
+//
+
+import Foundation
+
+@MainActor
+class UtxoCoinService: CoinService {}
